CHELSA

STNID

Unique station ID to link to GSOD data from GSODR

CHELSA_bio1_1979-2013_V1_1

Annual mean temperature [degree C]

CHELSA_bio2_1979-2013_V1_1

Mean diurnal range [degree C]

CHELSA_bio3_1979-2013_V1_1

Isothermality

CHELSA_bio4_1979-2013_V1_1

Temperature seasonality

CHELSA_bio5_1979-2013_V1_1

Maximum Temperature of warmest month [degree C]

CHELSA_bio6_1979-2013_V1_1

Minimum Temperature of coldest month [degree C]

CHELSA_bio7_1979-2013_V1_1

Temperature Annual Range [degree C]

CHELSA_bio8_1979-2013_V1_1

Mean Temperature of wettest quarter [degree C]

CHELSA_bio9_1979-2013_V1_1

Mean Temperature of driest quarter [degree C]

CHELSA_bio10_1979-2013_V1_1

Mean Temperature of warmest quarter [degree C]

CHELSA_bio11_1979-2013_V1_1

Mean Temperature of coldest quarter [degree C]

CHELSA_bio12_1979-2013_V1_1

Annual precipitation amount [mm]

CHELSA_bio13_1979-2013_V1_1

Precipitation of wettest month [mm]

CHELSA_bio14_1979-2013_V1_1

Precipitation of driest month [mm]

CHELSA_bio15_1979-2013_V1_1

Precipitation Seasonality

CHELSA_bio16_1979-2013_V1_1

Precipitation of wettest quarter [mm]

CHELSA_bio17_1979-2013_V1_1

Precipitation of driest quarter [mm]

CHELSA_bio18_1979-2013_V1_1

Precipitation of warmest quarter [mm]

CHELSA_bio19_1979-2013_V1_1

Precipitation of coldest quarter [mm]

CHELSA_prec_1_1979-2013

Mean January precipitation

CHELSA_prec_2_1979-2013

Mean February precipitation

CHELSA_prec_3_1979-2013

Mean March precipitation

CHELSA_prec_4_1979-2013

Mean April precipitation

CHELSA_prec_5_1979-2013

Mean May precipitation

CHELSA_prec_6_1979-2013

Mean June precipitation

CHELSA_prec_7_1979-2013

Mean July precipitation

CHELSA_prec_8_1979-2013

Mean August precipitation

CHELSA_prec_9_1979-2013

Mean September precipitation

CHELSA_prec_10_1979-2013

Mean October precipitation

CHELSA_prec_11_1979-2013

Mean November precipitation

CHELSA_prec_12_1979-2013

Mean December precipitation

CHELSA_prec_1979-2013_land

Mean annual precipitation

CHELSA_temp_1_1979-2013

Mean January temperature

CHELSA_temp_2_1979-2013

Mean February temperature

CHELSA_temp_3_1979-2013

Mean March temperature

CHELSA_temp_4_1979-2013

Mean April temperature

CHELSA_temp_5_1979-2013

Mean May temperature

CHELSA_temp_6_1979-2013

Mean June temperature

CHELSA_temp_7_1979-2013

Mean July temperature

CHELSA_temp_8_1979-2013

Mean August temperature

CHELSA_temp_9_1979-2013

Mean September temperature

CHELSA_temp_10_1979-2013

Mean October temperature

CHELSA_temp_11_1979-2013

Mean November temperature

CHELSA_temp_12_1979-2013

Mean December temperature

CHELSA_temp_1979-2013_land

Mean annual temperature

Source: R/CHELSA.R

CHELSA

STNID

Unique station ID to link to GSOD data from GSODR

CHELSA_bio1_1979-2013_V1_1

Annual mean temperature [degree C]

CHELSA_bio2_1979-2013_V1_1

Mean diurnal range [degree C]

CHELSA_bio3_1979-2013_V1_1

Isothermality

CHELSA_bio4_1979-2013_V1_1

Temperature seasonality

CHELSA_bio5_1979-2013_V1_1

Maximum Temperature of warmest month [degree C]

CHELSA_bio6_1979-2013_V1_1

Minimum Temperature of coldest month [degree C]

CHELSA_bio7_1979-2013_V1_1

Temperature Annual Range [degree C]

CHELSA_bio8_1979-2013_V1_1

Mean Temperature of wettest quarter [degree C]

CHELSA_bio9_1979-2013_V1_1

Mean Temperature of driest quarter [degree C]

CHELSA_bio10_1979-2013_V1_1

Mean Temperature of warmest quarter [degree C]

CHELSA_bio11_1979-2013_V1_1

Mean Temperature of coldest quarter [degree C]

CHELSA_bio12_1979-2013_V1_1

Annual precipitation amount [mm]

CHELSA_bio13_1979-2013_V1_1

Precipitation of wettest month [mm]

CHELSA_bio14_1979-2013_V1_1

Precipitation of driest month [mm]

CHELSA_bio15_1979-2013_V1_1

Precipitation Seasonality

CHELSA_bio16_1979-2013_V1_1

Precipitation of wettest quarter [mm]

CHELSA_bio17_1979-2013_V1_1

Precipitation of driest quarter [mm]

CHELSA_bio18_1979-2013_V1_1

Precipitation of warmest quarter [mm]

CHELSA_bio19_1979-2013_V1_1

Precipitation of coldest quarter [mm]

CHELSA_prec_1_1979-2013

Mean January precipitation

CHELSA_prec_2_1979-2013

Mean February precipitation

CHELSA_prec_3_1979-2013

Mean March precipitation

CHELSA_prec_4_1979-2013

Mean April precipitation

CHELSA_prec_5_1979-2013

Mean May precipitation

CHELSA_prec_6_1979-2013

Mean June precipitation

CHELSA_prec_7_1979-2013

Mean July precipitation

CHELSA_prec_8_1979-2013

Mean August precipitation

CHELSA_prec_9_1979-2013

Mean September precipitation

CHELSA_prec_10_1979-2013

Mean October precipitation

CHELSA_prec_11_1979-2013

Mean November precipitation

CHELSA_prec_12_1979-2013

Mean December precipitation

CHELSA_prec_1979-2013_land

Mean annual precipitation

CHELSA_temp_1_1979-2013

Mean January temperature

CHELSA_temp_2_1979-2013

Mean February temperature

CHELSA_temp_3_1979-2013

Mean March temperature

CHELSA_temp_4_1979-2013

Mean April temperature

CHELSA_temp_5_1979-2013

Mean May temperature

CHELSA_temp_6_1979-2013

Mean June temperature

CHELSA_temp_7_1979-2013

Mean July temperature

CHELSA_temp_8_1979-2013

Mean August temperature

CHELSA_temp_9_1979-2013

Mean September temperature

CHELSA_temp_10_1979-2013

Mean October temperature

CHELSA_temp_11_1979-2013

Mean November temperature

CHELSA_temp_12_1979-2013

Mean December temperature

CHELSA_temp_1979-2013_land

Mean annual temperature

CHELSA

Format

An object of class data.frame with 23927 rows and 46 columns.

Source

CHELSA climate layers (http://chelsa-climate.org/)

Note

CHELSA (climatic surfaces at 1 km resolution) is based on a quasi-mechanistic statistical downscaling of the ERA interim global circulation model (Karger et al. 2016). ESA's CCI-LC cloud probability monthly averages are based on the MODIS snow products (MOD10A2).

References

Karger, D. N., Conrad, O., Bohner, J., Kawohl, T., Kreft, H., Soria-Auza, R. W., et al. (2016). Climatologies at high resolution for the Earth land surface areas. arXiv preprint arXiv:1607.00217.