WorldClim_Clim

STNID

Unique station ID to link to GSOD data from GSODR

tmin1

Mean Minimum Temperature for January

tmin2

Mean Minimum Temperature for February

tmin3

Mean Minimum Temperature for March

tmin4

Mean Minimum Temperature for April

tmin5

Mean Minimum Temperature for May

tmin6

Mean Minimum Temperature for June

tmin7

Mean Minimum Temperature for July

tmin8

Mean Minimum Temperature for August

tmin9

Mean Minimum Temperature for September

tmin10

Mean Minimum Temperature for October

tmin11

Mean Minimum Temperature for November

tmin12

Mean Minimum Temperature for December

tmax1

Mean Maximum Temperature for January

tmax2

Mean Maximum Temperature for February

tmax3

Mean Maximum Temperature for March

tmax4

Mean Maximum Temperature for April

tmax5

Mean Maximum Temperature for May

tmax6

Mean Maximum Temperature for June

tmax7

Mean Maximum Temperature for July

tmax8

Mean Maximum Temperature for August

tmax9

Mean Maximum Temperature for September

tmax10

Mean Maximum Temperature for October

tmax11

Mean Maximum Temperature for November

tmax12

Mean Maximum Temperature for December

tmean1

Mean Temperature for January

tmean2

Mean Temperature for February

tmean3

Mean Temperature for March

tmean4

Mean Temperature for April

tmean5

Mean Temperature for May

tmean6

Mean Temperature for June

tmean7

Mean Temperature for July

tmean8

Mean Temperature for August

tmean9

Mean Temperature for September

tmean10

Mean Temperature for October

tmean11

Mean Temperature for November

tmean12

Mean Temperature for December

prec1

Mean Precipitation for January

prec2

Mean Precipitation for February

prec3

Mean Precipitation for March

prec4

Mean Precipitation for April

prec5

Mean Precipitation for May

prec6

Mean Precipitation for June

prec7

Mean Precipitation for July

prec8

Mean Precipitation for August

prec9

Mean Precipitation for September

prec10

Mean Precipitation for October

prec11

Mean Precipitation for November

prec12

Mean Precipitation for December

WorldClim_Clim

STNID

Unique station ID to link to GSOD data from GSODR

tmin1

Mean Minimum Temperature for January

tmin2

Mean Minimum Temperature for February

tmin3

Mean Minimum Temperature for March

tmin4

Mean Minimum Temperature for April

tmin5

Mean Minimum Temperature for May

tmin6

Mean Minimum Temperature for June

tmin7

Mean Minimum Temperature for July

tmin8

Mean Minimum Temperature for August

tmin9

Mean Minimum Temperature for September

tmin10

Mean Minimum Temperature for October

tmin11

Mean Minimum Temperature for November

tmin12

Mean Minimum Temperature for December

tmax1

Mean Maximum Temperature for January

tmax2

Mean Maximum Temperature for February

tmax3

Mean Maximum Temperature for March

tmax4

Mean Maximum Temperature for April

tmax5

Mean Maximum Temperature for May

tmax6

Mean Maximum Temperature for June

tmax7

Mean Maximum Temperature for July

tmax8

Mean Maximum Temperature for August

tmax9

Mean Maximum Temperature for September

tmax10

Mean Maximum Temperature for October

tmax11

Mean Maximum Temperature for November

tmax12

Mean Maximum Temperature for December

tmean1

Mean Temperature for January

tmean2

Mean Temperature for February

tmean3

Mean Temperature for March

tmean4

Mean Temperature for April

tmean5

Mean Temperature for May

tmean6

Mean Temperature for June

tmean7

Mean Temperature for July

tmean8

Mean Temperature for August

tmean9

Mean Temperature for September

tmean10

Mean Temperature for October

tmean11

Mean Temperature for November

tmean12

Mean Temperature for December

prec1

Mean Precipitation for January

prec2

Mean Precipitation for February

prec3

Mean Precipitation for March

prec4

Mean Precipitation for April

prec5

Mean Precipitation for May

prec6

Mean Precipitation for June

prec7

Mean Precipitation for July

prec8

Mean Precipitation for August

prec9

Mean Precipitation for September

prec10

Mean Precipitation for October

prec11

Mean Precipitation for November

prec12

Mean Precipitation for December

WorldClim_Clim

Format

An object of class data.frame with 26296 rows and 49 columns.

Source

WorlcClim Global Climate Data - Free climate data for ecological modeling and GIS http://www.worldclim.org/current.

WorldClim are freely available, average monthly climate data. Current conditions (interpolations of observed data, representative of 1960-1990) are freely available for download from http://www.worldclim.org/version1. Climatic elements include minimum, mean, maximum temperature and precipitation along with derived bioclimatic variables. WorldClim 1.4 (current conditions) are released under a Creative Commons Attribution-ShareAlike 4.0 International License http://creativecommons.org/licenses/by-sa/4.0/.

Note

Temperatures in GSODRdata are in degrees C, in the original WorldClim data they are in degrees C * 10.

References

Hijmans, R.J., S.E. Cameron, J.L. Parra, P.G. Jones and A. Jarvis, 2005. Very high resolution interpolated climate surfaces for global land areas. International Journal of Climatology 25: 1965-1978.